test: fix binWrapper tests
This change removes the `bin-wrapper` mock file, which we no longer need since we no longer use `bin-wrapper`.  Instead, it moves the mock for `src/binWrapper.js` into the test file, since this seems to be how we mock local modules (as opposed to dependencies).

This is not ideal, because now that `bin-wrapper` is _our_ code, we should not be mocking it at all.  It'd make sense to then create a mock for `download`.  As a result, `src/binWrapper.js` _has no test coverage_, which is bad.  If we want to keep it, we should consider copying (and likely converting to jest) `bin-wrapper`'s tests into our repo.